# test/Altinn.App.Api.Tests/OpenApi/swagger.yaml
openapi: 3.0.1
info:
  title: Altinn App Api
  version: v1
paths:
  '/{org}/{app}/instances/{instanceOwnerPartyId}/{instanceGuid}/actions':
    post:
      tags:
        - Actions
      parameters:
        - name: org
          in: path
          required: true
          schema:
            type: string
        - name: app
          in: path
          required: true
          schema:
            type: string
        - name: instanceOwnerPartyId
          in: path
          required: true
          schema:
            type: integer
            format: int32
        - name: instanceGuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: language
          in: query
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserActionRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/UserActionRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/UserActionRequest'
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/UserActionResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/UserActionResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/UserActionResponse'
        '400':
          description: Bad Request
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/xml:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/xml:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '409':
          description: Conflict
        '500':
          description: Server Error
        '401':
          description: Unauthorized
...
  '/{org}/{app}/instances/{instanceOwnerPartyId}/{instanceGuid}/payment':
    get:
      tags:
        - Payment
      parameters:
        - name: org
          in: path
          required: true
          schema:
            type: string
        - name: app
          in: path
          required: true
          schema:
            type: string
        - name: instanceOwnerPartyId
          in: path
          required: true
          schema:
            type: integer
            format: int32
        - name: instanceGuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: language
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/PaymentInformation'
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentInformation'
            text/json:
              schema:
                $ref: '#/components/schemas/PaymentInformation'
            application/xml:
              schema:
                $ref: '#/components/schemas/PaymentInformation'
            text/xml:
              schema:
                $ref: '#/components/schemas/PaymentInformation'
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/xml:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/xml:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  '/{org}/{app}/instances/{instanceOwnerPartyId}/{instanceGuid}/payment/order-details':
    get:
      tags:
        - Payment
      parameters:
        - name: org
          in: path
          required: true
          schema:
            type: string
        - name: app
          in: path
          required: true
          schema:
            type: string
        - name: instanceOwnerPartyId
          in: path
          required: true
          schema:
            type: integer
            format: int32
        - name: instanceGuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: language
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/OrderDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/OrderDetails'
            application/xml:
              schema:
                $ref: '#/components/schemas/OrderDetails'
            text/xml:
              schema:
                $ref: '#/components/schemas/OrderDetails'
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/xml:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/xml:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
...
components:
  schemas:
    ActionError:
      type: object
      properties:
        code:
          type: string
          nullable: true
        message:
          type: string
          nullable: true
        metadata:
          type: object
          additionalProperties:
            type: string
          nullable: true
      additionalProperties: false
...
    Address:
      type: object
      properties:
        name:
          type: string
          nullable: true
        addressLine1:
          type: string
          nullable: true
        addressLine2:
          type: string
          nullable: true
        postalCode:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        country:
          type: string
          nullable: true
      additionalProperties: false
...
    CardDetails:
      type: object
      properties:
        maskedPan:
          type: string
          nullable: true
        expiryDate:
          type: string
          nullable: true
      additionalProperties: false
    ClientAction:
      type: object
      properties:
        id:
          type: string
          nullable: true
        metadata:
          type: object
          additionalProperties: { }
          nullable: true
      additionalProperties: false
...
    InvoiceDetails:
      type: object
      properties:
        invoiceNumber:
          type: string
          nullable: true
      additionalProperties: false
...
    OrderDetails:
      type: object
      properties:
        paymentProcessorId:
          type: string
          nullable: true
        receiver:
          $ref: '#/components/schemas/PaymentReceiver'
        orderReference:
          type: string
          nullable: true
        currency:
          type: string
          nullable: true
        orderLines:
          type: array
          items:
            $ref: '#/components/schemas/PaymentOrderLine'
          nullable: true
        totalPriceExVat:
          type: number
          format: double
          readOnly: true
        totalVat:
          type: number
          format: double
          readOnly: true
        totalPriceIncVat:
          type: number
          format: double
          readOnly: true
      additionalProperties: false
...
    Payer:
      type: object
      properties:
        privatePerson:
          $ref: '#/components/schemas/PayerPrivatePerson'
        company:
          $ref: '#/components/schemas/PayerCompany'
        shippingAddress:
          $ref: '#/components/schemas/Address'
        billingAddress:
          $ref: '#/components/schemas/Address'
      additionalProperties: false
    PayerCompany:
      type: object
      properties:
        organisationNumber:
          type: string
          nullable: true
        name:
          type: string
          nullable: true
        contactPerson:
          $ref: '#/components/schemas/PayerPrivatePerson'
      additionalProperties: false
    PayerPrivatePerson:
      type: object
      properties:
        firstName:
          type: string
          nullable: true
        lastName:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        phoneNumber:
          $ref: '#/components/schemas/PhoneNumber'
      additionalProperties: false
    PaymentDetails:
      type: object
      properties:
        paymentId:
          type: string
          nullable: true
        redirectUrl:
          type: string
          nullable: true
        payer:
          $ref: '#/components/schemas/Payer'
        paymentType:
          type: string
          nullable: true
        paymentMethod:
          type: string
          nullable: true
        invoiceDetails:
          $ref: '#/components/schemas/InvoiceDetails'
        cardDetails:
          $ref: '#/components/schemas/CardDetails'
      additionalProperties: false
    PaymentInformation:
      type: object
      properties:
        taskId:
          type: string
          nullable: true
        status:
          $ref: '#/components/schemas/PaymentStatus'
        orderDetails:
          $ref: '#/components/schemas/OrderDetails'
        paymentDetails:
          $ref: '#/components/schemas/PaymentDetails'
      additionalProperties: false
    PaymentOrderLine:
      type: object
      properties:
        id:
          type: string
          nullable: true
        name:
          type: string
          nullable: true
        textResourceKey:
          type: string
          nullable: true
        priceExVat:
          type: number
          format: double
        quantity:
          type: integer
          format: int32
        vatPercent:
          type: number
          format: double
        unit:
          type: string
          nullable: true
      additionalProperties: false
    PaymentReceiver:
      type: object
      properties:
        organisationNumber:
          type: string
          nullable: true
        name:
          type: string
          nullable: true
        postalAddress:
          $ref: '#/components/schemas/Address'
      additionalProperties: false
    PaymentStatus:
      enum:
        - 0
        - 1
        - 2
        - 3
        - 4
        - 5
      type: integer
      format: int32
    PhoneNumber:
      type: object
      properties:
        prefix:
          type: string
          nullable: true
        number:
          type: string
          nullable: true
      additionalProperties: false
...
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        status:
          type: integer
          format: int32
          nullable: true
        detail:
          type: string
          nullable: true
        instance:
          type: string
          nullable: true
      additionalProperties: { }
...
    UserActionRequest:
      type: object
      properties:
        action:
          type: string
          nullable: true
        buttonId:
          type: string
          nullable: true
        metadata:
          type: object
          additionalProperties:
            type: string
          nullable: true
        ignoredValidators:
          type: array
          items:
            type: string
          nullable: true
      additionalProperties: false
    UserActionResponse:
      type: object
      properties:
        updatedDataModels:
          type: object
          additionalProperties: { }
          nullable: true
        updatedValidationIssues:
          type: object
          additionalProperties:
            type: object
            additionalProperties:
              type: array
              items:
                $ref: '#/components/schemas/ValidationIssue'
            nullable: true
          nullable: true
        clientActions:
          type: array
          items:
            $ref: '#/components/schemas/ClientAction'
          nullable: true
        error:
          $ref: '#/components/schemas/ActionError'
        redirectUrl:
          type: string
          format: uri
          nullable: true
      additionalProperties: false
    ValidationIssue:
      type: object
      properties:
        severity:
          $ref: '#/components/schemas/ValidationIssueSeverity'
        dataElementId:
          type: string
          nullable: true
        field:
          type: string
          nullable: true
        code:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        source:
          type: string
          nullable: true
        customTextKey:
          type: string
          nullable: true
        customTextParams:
          type: array
          items:
            type: string
          nullable: true
      additionalProperties: false
    ValidationIssueSeverity:
      enum:
        - 0
        - 1
        - 2
        - 3
        - 4
        - 5
      type: integer
      format: int32
...